refactor calibration actions and add pydantic model to action inputs
